Add additional MedCraft schemas
Base schema is a standard appsetttings validation with additional custom
application specific properties.

Local schema is intended only for local development of the application,
defining settings for simulating authentication in development only
local build configurations.

Due to tagged versioning where versions are only applied to schema URLs
upon publish, referencing other files of the same version is difficult.
As a workaround, simply mirror all validation in another "allOf" section
in a schema that should also reference the file. In this case, the local
schema has a copy of the base schema definitions, which for now will
just have to manually be kept synchronized